Split up run.py to gcp and aws runner
