+++ api/py/ai/chronon/repo/utils.py
@@ -1,7 +1,15 @@
+import json
 import logging
 import os
+import re
+import subprocess
 import time
+import xml.etree.ElementTree as ET
+
+from datetime import timedelta, datetime
 from enum import Enum
+
+from ai.chronon.repo.constants import *
 
 
 class DataprocJobType(Enum):
@@ -33,9 +41,9 @@ def retry_decorator(retries=3, backoff=20):
     return wrapper
 
 
-def get_environ_arg(env_name) -> str:
+def get_environ_arg(env_name, ignoreError=False) -> str:
     value = os.environ.get(env_name)
-    if not value:
+    if not value and not ignoreError:
         raise ValueError(f"Please set {env_name} environment variable")
     return value
 
@@ -46,3 +54,295 @@ def get_customer_id() -> str:
 
 def extract_filename_from_path(path):
     return path.split("/")[-1]
+
+def gen_final_args(mode, online_class="", online_jar="", ds="", conf=None, conf_type=None, start_ds=None, end_ds=None, override_conf_path=None, additional_user_args="", **kwargs):
+    base_args = MODE_ARGS[mode].format(
+        conf_path=override_conf_path if override_conf_path else conf,
+        ds=end_ds if end_ds else ds,
+        online_jar=online_jar,
+        online_class=online_class
+    )
+    base_args = base_args + f" --conf-type={conf_type} " if conf_type else base_args
+
+    override_start_partition_arg = (
+        "--start-partition-override=" + start_ds if start_ds else ""
+    )
+
+    additional_kwargs = " ".join(f"--{key.replace('_', '-')}={value}" for key, value in kwargs.items() if value)
+
+    final_args = " ".join([base_args, str(additional_user_args), override_start_partition_arg, additional_kwargs])
+
+    return final_args
+
+def check_call(cmd):
+    print("Running command: " + cmd)
+    return subprocess.check_call(cmd.split(), bufsize=0)
+
+
+def check_output(cmd):
+    print("Running command: " + cmd)
+    return subprocess.check_output(cmd.split(), bufsize=0).strip()
+
+
+def custom_json(conf):
+    """Extract the json stored in customJson for a conf."""
+    if conf.get("metaData", {}).get("customJson"):
+        return json.loads(conf["metaData"]["customJson"])
+    return {}
+
+
+
+def download_only_once(url, path, skip_download=False):
+    if skip_download:
+        print("Skipping download of " + path)
+        return
+    should_download = True
+    path = path.strip()
+    if os.path.exists(path):
+        content_output = check_output("curl -sI " + url).decode("utf-8")
+        content_length = re.search(
+            "(content-length:\\s)(\\d+)", content_output.lower())
+        remote_size = int(content_length.group().split()[-1])
+        local_size = int(check_output("wc -c " + path).split()[0])
+        print(
+            """Files sizes of {url} vs. {path}
+    Remote size: {remote_size}
+    Local size : {local_size}""".format(
+                **locals()
+            )
+        )
+        if local_size == remote_size:
+            print("Sizes match. Assuming it's already downloaded.")
+            should_download = False
+        if should_download:
+            print("Different file from remote at local: " +
+                  path + ". Re-downloading..")
+            check_call("curl {} -o {} --connect-timeout 10".format(url, path))
+    else:
+        print("No file at: " + path + ". Downloading..")
+        check_call("curl {} -o {} --connect-timeout 10".format(url, path))
+
+
+# NOTE: this is only for the open source chronon. For the internal zipline version, we have a different jar to download.
+@retry_decorator(retries=3, backoff=50)
+def download_jar(
+        version,
+        jar_type="uber",
+        release_tag=None,
+        spark_version="2.4.0",
+        skip_download=False,
+):
+    assert (spark_version in SUPPORTED_SPARK), (f"Received unsupported spark version {spark_version}. "
+                                                f"Supported spark versions are {SUPPORTED_SPARK}")
+    scala_version = SCALA_VERSION_FOR_SPARK[spark_version]
+    maven_url_prefix = os.environ.get("CHRONON_MAVEN_MIRROR_PREFIX", None)
+    default_url_prefix = (
+        "https://s01.oss.sonatype.org/service/local/repositories/public/content"
+    )
+    url_prefix = maven_url_prefix if maven_url_prefix else default_url_prefix
+    base_url = "{}/ai/chronon/spark_{}_{}".format(
+        url_prefix, jar_type, scala_version)
+    print("Downloading jar from url: " + base_url)
+    jar_path = os.environ.get("CHRONON_DRIVER_JAR", None)
+    if jar_path is None:
+        if version == "latest":
+            version = None
+        if version is None:
+            metadata_content = check_output(
+                "curl -s {}/maven-metadata.xml".format(base_url)
+            )
+            meta_tree = ET.fromstring(metadata_content)
+            versions = [
+                node.text
+                for node in meta_tree.findall("./versioning/versions/")
+                if re.search(
+                    r"^\d+\.\d+\.\d+{}$".format(
+                        r"\_{}\d*".format(release_tag) if release_tag else ""
+                    ),
+                    node.text,
+                )
+            ]
+            version = versions[-1]
+        jar_url = "{base_url}/{version}/spark_{jar_type}_{scala_version}-{version}-assembly.jar".format(
+            base_url=base_url,
+            version=version,
+            scala_version=scala_version,
+            jar_type=jar_type,
+        )
+        jar_path = os.path.join("/tmp", extract_filename_from_path(jar_url))
+        download_only_once(jar_url, jar_path, skip_download)
+    return jar_path
+
+
+def get_teams_json_file_path(repo_path):
+    return os.path.join(repo_path, "teams.json")
+
+
+def set_runtime_env(params):
+    """
+    Setting the runtime environment variables.
+    These are extracted from the common env, the team env and the common env.
+    In order to use the environment variables defined in the configs as overrides for the args in the cli this method
+    needs to be run before the runner and jar downloads.
+
+    The order of priority is:
+        - Environment variables existing already.
+        - Environment variables derived from args (like app_name)
+        - conf.metaData.modeToEnvMap for the mode (set on config)
+        - team's dev environment for each mode set on teams.json
+        - team's prod environment for each mode set on teams.json
+        - default team environment per context and mode set on teams.json
+        - Common Environment set in teams.json
+    """
+    environment = {
+        "common_env": {},
+        "conf_env": {},
+        "default_env": {},
+        "team_env": {},
+        "production_team_env": {},
+        "cli_args": {},
+    }
+    conf_type = None
+    # Normalize modes that are effectively replacement of each other (streaming/local-streaming/streaming-client)
+    effective_mode = params["mode"]
+    if effective_mode and "streaming" in effective_mode:
+        effective_mode = "streaming"
+    if params["repo"]:
+        teams_file = get_teams_json_file_path(params["repo"])
+        if os.path.exists(teams_file):
+            with open(teams_file, "r") as infile:
+                teams_json = json.load(infile)
+            # we should have a fallback if user wants to set to something else `default`
+            environment["common_env"] = teams_json.get("default", {}).get(
+                "common_env", {}
+            )
+            if params["conf"] and effective_mode:
+                try:
+                    _, conf_type, team, _ = params["conf"].split("/")[-4:]
+                except Exception as e:
+                    logging.error(
+                        "Invalid conf path: {}, please ensure to supply the relative path to zipline/ folder".format(
+                            params["conf"]
+                        )
+                    )
+                    raise e
+                if not team:
+                    team = "default"
+                # context is the environment in which the job is running, which is provided from the args,
+                # default to be dev.
+                if params["env"]:
+                    context = params["env"]
+                else:
+                    context = "dev"
+                logging.info(
+                    f"Context: {context} -- conf_type: {conf_type} -- team: {team}"
+                )
+                conf_path = os.path.join(params["repo"], params["conf"])
+                if os.path.isfile(conf_path):
+                    with open(conf_path, "r") as conf_file:
+                        conf_json = json.load(conf_file)
+                    environment["conf_env"] = (
+                        conf_json.get("metaData")
+                        .get("modeToEnvMap", {})
+                        .get(effective_mode, {})
+                    )
+                    # Load additional args used on backfill.
+                    if custom_json(conf_json) and effective_mode in [
+                        "backfill",
+                        "backfill-left",
+                        "backfill-final",
+                    ]:
+                        environment["conf_env"]["CHRONON_CONFIG_ADDITIONAL_ARGS"] = (
+                            " ".join(custom_json(conf_json).get(
+                                "additional_args", []))
+                        )
+                    environment["cli_args"]["APP_NAME"] = APP_NAME_TEMPLATE.format(
+                        mode=effective_mode,
+                        conf_type=conf_type,
+                        context=context,
+                        name=conf_json["metaData"]["name"],
+                    )
+                environment["team_env"] = (
+                    teams_json[team].get(context, {}).get(effective_mode, {})
+                )
+                # fall-back to prod env even in dev mode when dev env is undefined.
+                environment["production_team_env"] = (
+                    teams_json[team].get("production", {}).get(
+                        effective_mode, {})
+                )
+                # By default use production env.
+                environment["default_env"] = (
+                    teams_json.get("default", {})
+                    .get("production", {})
+                    .get(effective_mode, {})
+                )
+                environment["cli_args"]["CHRONON_CONF_PATH"] = conf_path
+    if params["app_name"]:
+        environment["cli_args"]["APP_NAME"] = params["app_name"]
+    else:
+        if not params["app_name"] and not environment["cli_args"].get("APP_NAME"):
+            # Provide basic app_name when no conf is defined.
+            # Modes like metadata-upload and metadata-export can rely on conf-type or folder rather than a conf.
+            environment["cli_args"]["APP_NAME"] = "_".join(
+                [
+                    k
+                    for k in
+                    [
+                        "chronon",
+                        conf_type,
+                        params["mode"].replace("-", "_") if params["mode"] else None,
+                    ]
+                    if k is not None
+                ]
+            )
+
+    # Adding these to make sure they are printed if provided by the environment.
+    environment["cli_args"]["CHRONON_DRIVER_JAR"] = params["chronon_jar"]
+    environment["cli_args"]["CHRONON_ONLINE_JAR"] = params["online_jar"]
+    environment["cli_args"]["CHRONON_ONLINE_CLASS"] = params["online_class"]
+    order = [
+        "conf_env",
+        "team_env",
+        "production_team_env",
+        "default_env",
+        "common_env",
+        "cli_args",
+    ]
+    print("Setting env variables:")
+    for key in os.environ:
+        if any([key in environment[set_key] for set_key in order]):
+            print(f"From <environment> found {key}={os.environ[key]}")
+    for set_key in order:
+        for key, value in environment[set_key].items():
+            if key not in os.environ and value is not None:
+                print(f"From <{set_key}> setting {key}={value}")
+                os.environ[key] = value
+
+
+def split_date_range(start_date, end_date, parallelism):
+    start_date = datetime.strptime(start_date, "%Y-%m-%d")
+    end_date = datetime.strptime(end_date, "%Y-%m-%d")
+    if start_date > end_date:
+        raise ValueError("Start date should be earlier than end date")
+    total_days = (
+                         end_date - start_date
+                 ).days + 1  # +1 to include the end_date in the range
+
+    # Check if parallelism is greater than total_days
+    if parallelism > total_days:
+        raise ValueError(
+            "Parallelism should be less than or equal to total days")
+
+    split_size = total_days // parallelism
+    date_ranges = []
+
+    for i in range(parallelism):
+        split_start = start_date + timedelta(days=i * split_size)
+        if i == parallelism - 1:
+            split_end = end_date
+        else:
+            split_end = split_start + timedelta(days=split_size - 1)
+        date_ranges.append(
+            (split_start.strftime("%Y-%m-%d"), split_end.strftime("%Y-%m-%d"))
+        )
+    return date_ranges
