+++ api/py/requirements/base.in
@@ -6,3 +6,4 @@ pyspark==3.5.4
 sqlglot
 crcmod==1.7
 glom
+boto3
