+++ api/py/requirements/base.txt
@@ -5,7 +5,12 @@
 #
 #    pip-compile-multi
 #
-
+boto3==1.37.6
+    # via -r requirements/base.in
+botocore==1.37.6
+    # via
+    #   boto3
+    #   s3transfer
 cachetools==5.5.0
     # via google-auth
 charset-normalizer==3.4.1
@@ -37,6 +42,10 @@ googleapis-common-protos==1.66.0
     # via google-api-core
 idna==3.10
     # via requests
+jmespath==1.0.1
+    # via
+    #   boto3
+    #   botocore
 proto-plus==1.25.0
     # via google-api-core
 protobuf==5.29.3
@@ -50,15 +59,21 @@ pyasn1==0.6.1
     #   rsa
 pyasn1-modules==0.4.1
     # via google-auth
+python-dateutil==2.9.0.post0
+    # via botocore
 requests==2.32.3
     # via
     #   google-api-core
     #   google-cloud-storage
 rsa==4.9
     # via google-auth
+s3transfer==0.11.4
+    # via boto3
 six==1.17.0
     # via thrift
 thrift==0.21.0
     # via -r requirements/base.in
 urllib3==2.3.0
-    # via requests
+    # via
+    #   botocore
+    #   requests
