+++ api/py/test/sample/teams.json
@@ -16,7 +16,8 @@
       "GCP_PROJECT_ID": "canary-443022",
       "GCP_REGION": "us-central1",
       "GCP_DATAPROC_CLUSTER_NAME": "zipline-canary-cluster",
-      "GCP_BIGTABLE_INSTANCE_ID": "zipline-canary-instance"
+      "GCP_BIGTABLE_INSTANCE_ID": "zipline-canary-instance",
+      "CLOUD_PROVIDER": "GCP"
     },
     "production": {
       "backfill": {
